Make mpi depend on mpi-derive and re-export derives
