+++ Cargo.toml
@@ -28,6 +28,7 @@ conv = "0.3"
 libffi = { version = "0.8.0", optional = true }
 mpi-sys = { path = "mpi-sys", version = "0.2" }
 smallvec = "1.0.0"
+mpi-derive = { path = "mpi-derive" }
 
 [dev-dependencies]
 memoffset = "0.5"
