+++ src/lib.rs
@@ -121,6 +121,9 @@
 use std::mem::MaybeUninit;
 use std::os::raw::c_int;
 
+// Re-export derives
+pub use mpi_derive::*;
+
 /// The raw C language MPI API
 ///
 /// Documented in the [Message Passing Interface specification][spec]
